+++ StepResponse/SimulationModel/IParametrable.cs
@@ -7,18 +7,18 @@ namespace StepResponse.SimulationModel
         /// <summary>
         /// Return parameters for UI/export
         /// </summary>
-        public abstract Dictionary<string, float> GetParameters();
+        public abstract Dictionary<string, double> GetParameters();
 
         /// <summary>
         /// Get parameter by name.
         /// </summary>
-        public abstract bool GetParameter(string param, out float value);
+        public abstract bool GetParameter(string param, out double value);
 
         /// <summary>
         /// Set parameter by name.
         /// </summary>
-        public abstract bool SetParameter(string param, float value);
+        public abstract bool SetParameter(string param, double value);
 
-        public abstract bool IsValidValue(string param, float value);
+        public abstract bool IsValidValue(string param, double value);
     }
 }
